requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymilvus==2.6.5
milvus-lite==2.5.1
sentence-transformers==5.2.0
torch==2.9.1
transformers==4.57.3
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
pandas==2.3.3
pydantic==2.5.0
huggingface-hub==0.36.0
safetensors==0.7.0
tokenizers==0.22.1
tqdm==4.67.1
protobuf==6.33.2
grpcio==1.76.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pyyaml==6.0.1
PyJWT==2.8.0
requests>=2.31.0
python-dateutil>=2.8.2
